[Builder] Split builder into a base/kaniko/__init__ package (pure move)
Behaviour-preserving relocation only, no logic change: the single
services/api/utils/builder.py module becomes a package split into three layers
that import only downward -

- base.py    engine-agnostic building blocks (dockerfile render, arg parsing,
             image-name / requirements / builder-env resolution)
- kaniko.py  the Kaniko pieces (make_kaniko_pod, ECR + source-fetch + arg
             helpers), imports base
- __init__.py the entry points (build_image, build_runtime), imports base + kaniko
             and re-exports the public surface

Every function moves verbatim; the only edits are the mechanical cross-file
plumbing (build_image now calls base.make_dockerfile / kaniko.make_kaniko_pod)
and repointing the moved-symbol references in the tests. No behaviour change.
